Move remaining configure for features and elements
